Updating DumpSolutions to output different quantities.
Quanties output are now based on the parameters specified
in the input.in file.

DumpSolutions also now recognises a new code: -2. This tells
it that the last orbit is complete, and now it must dump
orbitally averaged properties as well (lines 713-724).

Additional changes are pending for dissipation quantities.

Now dissFile is declared within a local "if" statement.